Fix for java.lang.IllegalArgumentException : Wrong FS
The Configuration object used to get the FileSystem object was set in
the Map Task constructor. But at constructor time, the configuration
object points to the default configuration, not the actual configuration
used by the job (i.e. when the job is submitted).  This configuration
could be different from the default configuration and is specified in
the job.xml file  for the job.  The actual configuration is sent by the
MapRunner using callback for configure method and should be used to get
the actual job configuration.