Enable default support for hibench.yarn.executor.num
Current implementation assumes the "hibench.yarn.executor.num", that
is, "--num-executors" option, is not empty. The "--num-executors" option
specifies the container num in the whole YARN cluster and it is usually
kept default when submit a application in practice, to make the
application's executors specified by the cluster. Actually, the community
has removed this option in the Spark doc, See
http://spark.apache.org/docs/latest/running-on-yarn.html.
What's more, For porpose of benchmark, we hope the application can
 use up all of the resouces of the cluster.

Changes:
 * bin/functions/workload-functions.sh: append the --num-executors
   option only if the hibench.yarn.executor.num is not none.